Search Bar: Factor out the Filter Suggestions to their own component
Move the html and its relevant logic to a dedicated component
